% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{ffp_doTransform}
\alias{ffp_doTransform}
\title{Static function to performs biexponential transform on all parameters except SSC-W
This is verbatim what Wade was using originally}
\usage{
ffp_doTransform(f, cols = c(1:5, 7:13), method = c("biexp", "log",
  "linear"), fac = 5.4/262143)
}
\description{
Static function to performs biexponential transform on all parameters except SSC-W
This is verbatim what Wade was using originally
}
