% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{ffp_initialize}
\alias{ffp_initialize}
\title{Construct a flowFramePlus}
\usage{
ffp_initialize(self, private, ff = NA, txlist = NA, plist = c("FSC-H",
  "SSC-H"))
}
\arguments{
\item{self}{this}

\item{private}{this$private}

\item{ff}{flowFrame object or FCS file path}

\item{txlist}{Which columns to transform and how}

\item{plist}{Which columns to plot on the x and y axis by default (optional)}
}
\description{
Construct a flowFramePlus
}
\keyword{internal}
