% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ffp_plot}
\alias{ffp_plot}
\title{Plot a flowFramePlus use base R (pplot)}
\usage{
ffp_plot(self, private, plist = self$plist, blueBackground = FALSE,
  showZero = TRUE, nbin = 501, bandwidth = 0.001,
  cr = blob_color(blueBackground = FALSE), col = "black", nrpoints = 0,
  instrument = c("diva", "influx"), tx = self$plotScales$x,
  ty = self$plotScales$y, plotaxt = T, ticksize = 1, ...)
}
\arguments{
\item{self}{this}

\item{private}{this$private}

\item{plist}{Which columns to plot on the x and y axes, can override settings made at class instantiation}

\item{tx}{Scale to apply to x axis}

\item{ty}{Scale to apply to y axis}
}
\description{
Plot a flowFramePlus use base R (pplot)
}
\keyword{internal}
