% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{ffp_transformCols}
\alias{ffp_transformCols}
\title{Member function performs biexponential transform on all parameters except SSC-W}
\usage{
ffp_transformCols(self, private, cols, method)
}
\arguments{
\item{private}{this$private}

\item{cols}{Scale to apply to x axis}

\item{method}{transformation method}

\item{f}{flowFrame}
}
\value{
a transformed flowFrame
}
\description{
Member function performs biexponential transform on all parameters except SSC-W
}
\details{
This function performs biexponential transform on all parameters except SSC-W

Jitter is set to false for biexponential to get reproducible results (otherwise would result in random #)
}
\keyword{internal}
