% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowFramePlus.R
\docType{data}
\name{flowFramePlus}
\alias{flowFramePlus}
\title{flowFramePlus}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
flowFramePlus
}
\description{
flowFramePlus holds transformalization and scale attributes. It is essentially a small
wrapper around the \code{flowFrame} with some logical defaults.
}
\section{Usage}{

\preformatted{p <- flowFramePlus$new(transformlist = NA, plist = c("FSC-H","SSC-H"))}
\preformatted{p <- flowFramePlus$plot()}
}

\keyword{datasets}
