% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSetPlus.R
\name{fsp_initialize}
\alias{fsp_initialize}
\title{Construct a flowSetPlus}
\usage{
fsp_initialize(self, private, frames = NA, txlist = NA, plist = c("FSC-H",
  "SSC-H"))
}
\arguments{
\item{self}{this}

\item{private}{this$private}

\item{frames}{flowFrame object(s) or FCS file path(s)}

\item{txlist}{Which columns to transform and how}

\item{plist}{Which columns to plot on the x and y axis by default (optional)}
}
\description{
Construct a flowSetPlus
}
\keyword{internal}
