% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{fsp_plot_classic}
\alias{fsp_plot_classic}
\title{Plot a flowSetPlus use base R (pplot)}
\usage{
fsp_plot_classic(self, private, main = NULL, ...)
}
\arguments{
\item{self}{this}

\item{private}{this$private}

\item{main}{Plot titles}

\item{plist}{Which columns to plot on the x and y axes, can override settings made at class instantiation}

\item{tx}{Scale to apply to x axis}

\item{ty}{Scale to apply to y axis}
}
\description{
Plot a flowSetPlus use base R (pplot)
}
\keyword{internal}
