% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{fsp_plot_flowviz}
\alias{fsp_plot_flowviz}
\title{Plot a flowSetPlus using flowviz xyplot}
\usage{
fsp_plot_flowviz(self, private, ...)
}
\arguments{
\item{self}{this}

\item{private}{this$private}
}
\description{
Plot a flowSetPlus using flowviz xyplot
}
\keyword{internal}
