% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{ffp_doTransform}
\alias{ffp_doTransform}
\title{This function performs biexponential transform on all parameters except SSC-W}
\usage{
ffp_doTransform(self, private, f, cols = c(1:5, 7:13), method = c("biexp",
  "log", "linear"), fac = 5.4/262143)
}
\arguments{
\item{private}{this$private}

\item{f}{flowFrame}

\item{cols}{Scale to apply to x axis}

\item{method}{transformation method}
}
\value{
a transformed flowFrame
}
\description{
This function performs biexponential transform on all parameters except SSC-W
}
\details{
This function performs biexponential transform on all parameters except SSC-W

Jitter is set to false for biexponential to get reproducible results (otherwise would result in random #)
}
\keyword{internal}
