% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowFramePlus.R, R/flowSetPlus.R
\name{flowFramePlus}
\alias{flowFramePlus}
\alias{flowFramePlus}
\title{flowFramePlus}
\description{
flowFramePlus holds transformalization and scale attributes. It is essentially a small
wrapper around \code{flowFrame} with some logical defaults.

flowSetPlus holds one for more \code{flowFramePlus} objects with some logical defaults.
}
\section{Usage}{

\preformatted{p <- flowFramePlus$new("/projects/AML/data/AML/AML 1/AML 1 De Novo Panel_Tube_001.fcs", plist = c("FSC-H","SSC-H"))}
\preformatted{p <- flowFramePlus$plot()}


\preformatted{p <- flowSetPlus$new("/projects/AML/data/AML/AML 1/AML 1 De Novo Panel_Tube_001.fcs", plist = c("FSC-H","SSC-H"))}
\preformatted{p <- flowSetPlus$plot()}
}

\section{Arguments}{

\describe{
  \item{ff}{A \code{flowFrame} object or filepath of an FCS file.}
  \item{txlist}{List, specify transformation to be used on each column index.
  For example \code{txlist=list("linear"=c(1:4),"biexp"=c(5:12))}}
}


\describe{
  \item{ff}{A \code{flowFrame} object or filepath of an FCS file.}
  \item{txlist}{List, specify transformation to be used on each column index.
  For example \code{txlist=list("linear"=c(1:4),"biexp"=c(5:12))}}
}
}

\section{Details}{

\code{$new()} starts a new flowFramePlus object, either from an existing flowFrame object or
an FCS filepath

You can specify which columns to transform using the following syntax:
\preformatted{p <- flowFramePlus$new(x,txlist=list("linear"=c(1:4),"biexp"=c(5:12)))}


\code{$new()} starts a new flowFramePlus object, either from an existing flowFrame object or
an FCS filepath

You can specify which columns to transform using the following syntax:
\preformatted{p <- flowFramePlus$new(x,txlist=list("linear"=c(1:4),"biexp"=c(5:12)))}
}

\examples{
ffp<-flowFramePlus$new("/projects/AML/data/AML/AML 1/AML 1 De Novo Panel_Tube_001.fcs")
ffp$plot()
files<-dir(path="/projects/AML/data/AML/",pattern = "*.fcs",recursive = TRUE,include.dirs = TRUE,full.names = TRUE)
fsp<-flowSetPlus$new(files)
fsp$plot(plist=c("FSC-A","Horizon V500-A"))
}
